requirements.txt
aiogram>=3.4.0
python-dotenv
asyncpg
fastapi
uvicorn
jinja2